prepare rc, formatting
use brackets for single line if expressions
remove unused OfflineHandler::getOfflineValuesLeft function
fix missing includes from deleted common.h file
delete copy assignment operator of PrivacyIDEA
set new version

+++ CppClientCore/CppClientCore/OfflineData.cpp
@@ -134,7 +134,7 @@ int OfflineData::getLowestKey()
 	return lowestKey;
 }
 
-int OfflineData::getOfflineOTPsLeft() noexcept
+size_t OfflineData::getOfflineOTPsLeft() noexcept
 {
 	return offlineOTPs.size();
 }
